part4.d Token authentication

+++ part4/blog-list/controllers/blogs.js
@@ -1,20 +1,54 @@
 const blogsRouter = require("express").Router();
 const Blog = require("../models/blog");
+const middleware = require("../utils/middleware");
 
 blogsRouter.get("/", async (request, response) => {
-  const blogs = await Blog.find({});
+  const blogs = await Blog.find({}).populate("user", {
+    username: 1,
+    name: 1,
+  });
   response.json(blogs);
 });
 
-blogsRouter.post("/", async (request, response) => {
-  const blog = new Blog(request.body);
+blogsRouter.post("/", middleware.userExtractor, async (request, response) => {
+  const { title, author, url, likes } = request.body;
+  const user = request.user;
+
+  const blog = new Blog({
+    title,
+    author,
+    url,
+    likes,
+    user: user.id,
+  });
 
   const result = await blog.save();
+  user.blogs = user.blogs.concat(blog._id);
+  await user.save();
   response.status(201).json(result);
 });
 
-blogsRouter.delete("/:id", async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id);
+blogsRouter.delete(
+  "/:id",
+  middleware.userExtractor,
+  async (request, response) => {
+    const user = request.user;
+    const blog = await Blog.findById(request.params.id);
+
+    if (blog.user.toString() !== user.id) {
+      return response.status(401).json({ error: "invalid used" });
+    }
+
+    user.blogs = user.blogs.filter((b) => b.toString() !== blog.id);
+    await user.save();
+
+    await Blog.findByIdAndDelete(request.params.id);
+    response.status(204).end();
+  }
+);
+
+blogsRouter.delete("/", async (request, response) => {
+  await Blog.deleteMany({});
   response.status(204).end();
 });
